Upgraded image preview
It no longer removes the <input> element from the page but changes the value of the <input> element and changes the src of the already existing image. Incidentally I should make a default image header that will show up as part of the blog card on the blog posts page.