File upload upgrade
The file upload script now checks the blog_images directory to see how many folders there are (corresponding to the amount of pages) and creates a new folder and puts all the submitted files into the folder. It also sets the permissions correctly. Finally, the actual blog creation page has been restructured slightly to help with formatting and eventually saving the contents of the form as a html document.