JS-PHP communication
After messing around with the fetch API I managed to create a system that communicates how many pages there with javascript. Given that JS doesn't have access to the filesystem I just use PHP to log how many pages there are in a JSON file and have Javascript call a fetch on that file. After this is the long arduous task of writing the contents of the blog space into a HTML file and saving it.